Initial commit
The bsdmaps package is a nodejs application for collaborating with the
creation of possible attendance maps for the Beaverton School District
high schools.

It provides a data entry app for extracting the data provided in the
various PDF maps as well as the start of a solution app for making
map proposals. Only the data entry app is complete.

The concept and initial implementation are the work of Brad Larson. The
overlays were added by Darren Hart. Initial repository creation and
publication were done by Darren Hart.

Signed-off-by: Darren Hart <dvhart@infradead.org>