Refactor results to scope update
The charts were not updating after a load.

Ultimately this was due to the results to scope update being performed
in three places and the one after the load was not updating all the
scope.data fields. Refactor this so the calculation is done in a
function and called from each of the 3 call sites.

Confusing the issue was the fact that the $scope.apply() was being
called from synchronous places that resulted in apply() throwing its
already running error. Correct these by remove the calls to apply()
where these are triggered.

Charts now update after map load.
Other features continue to work while the apply() exceptions are no
longer thrown.

Signed-off-by: Darren Hart <dvhart@infradead.org>