Add drag to paint feature
Clicking each grid to assign it to a high school was tedious. This
change allows you to click and drag over grids or elementary school
regions to assign it to high schools, Much faster and cooler! Statistics
are computed when the mouse button is released if grids have changed.

Adapted from Brad Larsons original implementation.

Signed-off-by: Darren Hart <dvhart@infradead.org>