Allow for modifying FRL data without reordering the grids in the DB
Previously, the db did not hold the flr yellow dots, just the blue.
Editing the database would reorder the grids.  This currently breaks the
solution integration with the map.  With this commit, we are able to add
both blue and yellow dots from the FLR map to the database without
changing the grid location or ID.

Applied from Brad Larson's original commit.

Signed-off-by: Darren Hart <dvhart@infradead.org>