Prevent duplicate solution names with identical maps
Repeat solutions were being added to the solution database because there
was no solution checking and no ability for users to delete solutions.
This change adds a check for a solution name and unique solution entry
for the solution name. It enhances the save message to indicate the
type of failure that occurred.

Adapted from Brad Larson's original commit.

Signed-off-by: Darren Hart <dvhart@infradead.org>