Assign by distance
There was no good way for users to know if assigning a grid to one
school or another would reduce the travel distance although this
information was in the grid data.  Add the ability to assign to closest
high school so this can be achieved directly. Assigning each student to
the closest high school would result in 23471.03 miles and 3558
transitions.

Adapted from Brad Larson's original commit.

Signed-off-by: Darren Hart <dvhart@infradead.org>