Add Google Map API Markers to High School Locations
Make the high school locations more visible by adding Google Map Markers
at each location. We already know the location from the schoolData
structure, so the addition of the markers is a trivial for loop during
map initialization.

Use the first character of the school name in the pin icon. Add a
tooltip with the full school name. Make the pins not clickable so they
pass the click through to the underlying grid code.

Signed-off-by: Darren Hart <dvhart@infradead.org>