Ensure stats get updated after dragging
When dragging applies proposed high school, if the user releases the
drag somewhere other than a map.data object (grid code), so in the map
area, or on the web page, the stats will not update as there is no
listener for that event.

We can't use dragend because we disabled the drag events to prevent the
map from panning.

Instead, update the stats after each mouseover that applies the proposed
high school.

Signed-off-by: Darren Hart <dvhart@infradead.org>