Separate school types in schoolData
The addition of middle and elementrary school data to the schoolData
schools array broke logic in the app that assumed it contained only
highschools, resulting in a NaN in the stats total capacity metric.

It also added unnecessary overhead as all use cases are looping through
one type of school. Rather than loop through them all, separate them
into seperate arrays by type, and remove the type attribute, as it is
now implied.

Signed-off-by: Darren Hart <dvhart@infradead.org>