Add separate route section to database
Adding the route section to the grid database would grow the size of the grid database by several times.   This would slow down loading the solution explorer. Because the routes to school are not used by the solution explorer, direclty but only to comptue the distance and accident rate to schools, I will place the route in a seprate route location of the database.  This change gives read/write access to this section.