Enrollment: Resolve differences between BSD forecast & error analysis
Error analysis 2020 forecast was higher than the BSD 2020
forecast.  This delivery corrects a few minor errors and
adds instrumentation to identify the error.  The source
of the error is the enrollment data used for the BSD 2020
forecast is lower than the enrollment reported to the
Oregon Department of Education.  I am sticking to the
Oregon Department of Education data since it comes from
a consistent time of year and hopefully a consistent reporting
method.

Signed-off-by: Brad Larson <bhlarson@gmail.com>
Signed-off-by: Darren Hart <dvhart@infradead.org>